Lobby wire honesty: family at full byte range, names clamped at the writers (#218)
One contract violation behind four CI-only test failures and the TSan
lane's build error, all introduced by C1's wire hardening (cee5f4a1):
the serializers must be lossless for every message an honest peer
sends, and the read-side clamps broke that at both ends.

Family (the clang -Wtautological-constant-out-of-range-compare build
error): the wire carries an unsigned family byte — every value 0..255
indexes a real loader slot (NUM_FAMILY_SLOTS == 256) — but
LobbyCharacterData/InitialSetupGuyData stored it in int8_t. Clang is
right that `family >= NUM_FAMILY_SLOTS` can never fire on that type;
worse, wire families 128..255 aliased to negatives, so the `< 0` arm
silently ate real class-pack slots, and the typed in-process channel
(which skips the wire readers entirely) had no live upper bound. The
field is now int16_t, the reader widens the byte as-is, and the merge
compares at int width — both arms reachable, pinned by unit tests
(family 300 and -1 dropped, 200 kept, wire byte 200 decodes as +200).
A clang sweep over gameplay/server/core found no other instance of the
warning class.

Names (the four "round-trip mismatch" failures — two CursesNetwork
tests, GameLoop ally-mode, NetTransportInProcess last-survivor): C1
added read-side truncation for guy names (11 chars) and player/company
labels (40). Under VALIDATE_SERIALIZATION=ON — set in exactly the
failing CI lanes and OFF in a default local build, which is why the
local suite stayed green — every typed in-process message round-trips
through the serializers and must compare equal, so any roster carrying
a fixture name longer than the clamp now failed in transit. (Not the
suspected setup_generation hole: that v13 field is carried
symmetrically by the one serializer and default-initializes to 0; the
round-trip fixture now pins it nonzero regardless, so a decoder that
skipped the field can never pass by zero-luck.)

Fixed at the writers: shared clamp_lobby_guy_name /
clamp_lobby_player_label helpers in lobby_state.h, applied by the three
slot builders (curses, SDL network client, local picker lobby),
collect_initial_setup_guys (the world guy keeps its full name; only the
wire copy clamps), the player-label builders, and — for typed
in-process rosters the readers never see — the LobbyServer merge. The
writer clamps are pinned red-first at all three layers (curses builder,
server collector, merge), and the wire readers stay as the
crafted-payload backstop.

Test adjustments, none weakening: fixture guys named wider than the
wire width where the full name is asserted downstream are renamed to
fit ("Advanced Guest" -> "AdvGuest", "Yellow Middle" -> "Yellow Mid");
the curses internal check now pins the clamped "Curses Sold" the
builder must produce; and the oversize-broadcast guard test lost its
vehicle (names can no longer assemble an oversize state), so it drives
the guard through the one string sanitize leaves unbounded — a
host-authored campaign id — and asserts the state really went oversize
before the skipped send.

+++ include/openglad/gameplay/net_transport.h
@@ -236,7 +236,9 @@ struct HelloMessage {
 struct InitialSetupGuyData {
     std::int32_t guy_id = 0;
     std::string name;
-    std::int8_t family = 0;
+    // Wire byte 0..255; wider than the byte for the same reason as
+    // LobbyCharacterData::family (see lobby_state.h).
+    std::int16_t family = 0;
     std::int16_t strength = 0;
     std::int16_t dexterity = 0;
     std::int16_t constitution = 0;
